+++ constants/panels/panels.go
@@ -1,9 +1,9 @@
 package panels
 
 const (
-	SubmissionsPanel = "submissions"
-	InfoPanel        = "info"
-	KeymapsPanel     = "keymaps"
-	SettingsPanel    = "settings"
-	ModalPanel       = "modal"
+	SubmissionsPanel       = "submissions"
+	InfoPanel              = "info"
+	KeymapsPanel           = "keymaps"
+	SettingsPanel          = "settings"
+	AddCustomFavoritePanel = "custom_favorite"
 )