+++ constants/state/state.go
@@ -1,0 +1,7 @@
+package state
+
+const (
+	OnSubmissionPage = 0
+	Offline          = 1
+	OnHelpScreen     = 2
+)